+++ modules/coordination_module/Cargo.toml
@@ -10,3 +10,5 @@ tokio = { version = "1", features = ["full"] }
 colored = "2.0"
 json_comparator = { path = "json_comparator" }
 fingerprint = { path = "fingerprint" }
+json = { path = "json" }
+
