+++ modules/coordination_module/fingerPrint/src/check/mod.rs
@@ -1,0 +1,1 @@
+pub mod check_hash;
