+++ modules/coordination_module/fingerPrint/src/create/mod.rs
@@ -1,0 +1,1 @@
+pub mod create_hash;
