+++ modules/coordination_module/fingerPrint/src/insert/mod.rs
@@ -1,0 +1,1 @@
+pub mod insert_hash;
