+++ modules/coordination_module/json/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "json"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+tokio = { version = "1", features = ["full"] }
