+++ modules/coordination_module/json/json_objects/json1.json
@@ -1,0 +1,1 @@
+{"game": "World of Warcraft", "character": "Thrall", "ability": "Earthquake (Ultimate)", "damage": 1000, "type": "AoE"}
