+++ modules/coordination_module/json/json_objects/json5.json
@@ -1,0 +1,1 @@
+{"game": "World of Warcraft", "character": "Jaina", "ability": "Frostbolt"}
