+++ modules/coordination_module/json/src/mod.rs
@@ -1,0 +1,1 @@
+pub use crate::json::load_json_objects;
