+++ modules/coordination_module/src/mod.rs
@@ -1,0 +1,3 @@
+pub mod json;
+pub mod json_comparator;
+pub mod fingerprint;
